Add Dockerfile, update CMAKELists and setup.py
* add a dockerfile and build_wheels.sh for easier development
* add pybind11 and CMakeModules, and remove unnecessary statements in CMakeLists.txt
* add format files (.clang-format and pyproject.toml)
* add a basic unittest
* move Planner to a separate file
