+++ data/panda/franka_description/meshes/collision/finger.stl.convex.stl
