+++ data/panda/franka_description/meshes/collision/hand.stl
