+++ data/panda/franka_description/meshes/collision/hand.stl.convex.stl
