+++ data/panda/franka_description/meshes/collision/link0.stl
