+++ data/panda/franka_description/meshes/collision/link1.stl
