+++ data/panda/franka_description/meshes/collision/link2.stl
