+++ data/panda/franka_description/meshes/collision/link3.stl
