+++ data/panda/franka_description/meshes/collision/link4.stl
