+++ data/panda/franka_description/meshes/collision/link6.stl
