+++ data/panda/franka_description/meshes/collision/link7.stl.convex.stl
