+++ data/panda/franka_description/meshes/visual/finger.dae
