+++ data/panda/franka_description/meshes/visual/hand.dae
