+++ data/panda/franka_description/meshes/visual/link0.dae
