+++ data/panda/franka_description/meshes/visual/link1.dae
