+++ data/panda/franka_description/meshes/visual/link2.dae
