+++ data/panda/franka_description/meshes/visual/link3.dae
