+++ data/panda/franka_description/meshes/visual/link4.dae
