+++ data/panda/franka_description/meshes/visual/link6.dae
