+++ data/panda/franka_description/meshes/visual/link7.dae
