+++ dev/DEV.md
@@ -1,0 +1,18 @@
+# Developer Notes
+
+## Getting Started
+
+We have provided a docker image `mplib-build` based on pypa base image, to compile and package the library.
+Most dependencies have been installed.
+The docker file can be found [here](../docker/Dockerfile).
+
+## VSCode
+
+
+## FAQ
+
+<details>
+<summary>How to create the wheel?</summary>
+
+`bash` d
+</details>
