+++ pyproject.toml
@@ -1,0 +1,2 @@
+[tool.black]
+line_length = 79
