+++ setup.py
@@ -6,14 +6,18 @@ import subprocess
 from setuptools import setup, Extension
 from setuptools.command.build_ext import build_ext
 
+
 class CMakeExtension(Extension):
     def __init__(self, name, sourcedir=""):
         Extension.__init__(self, name, sources=[])
         self.sourcedir = os.path.abspath(sourcedir)
 
+
 class CMakeBuild(build_ext):
     def build_extension(self, ext):
-        extdir = os.path.abspath(os.path.dirname(self.get_ext_fullpath(ext.name)))
+        extdir = os.path.abspath(
+            os.path.dirname(self.get_ext_fullpath(ext.name))
+        )
 
         if not extdir.endswith(os.path.sep):
             extdir += os.path.sep
@@ -26,14 +30,16 @@ class CMakeBuild(build_ext):
         cmake_args = [
             "-DCMAKE_LIBRARY_OUTPUT_DIRECTORY={}".format(extdir),
             "-DPYTHON_EXECUTABLE={}".format(sys.executable),
-            "-DCMAKE_BUILD_TYPE={}".format(cfg),  # not used on MSVC, but no harm
+            "-DCMAKE_BUILD_TYPE={}".format(
+                cfg
+            ),  # not used on MSVC, but no harm
         ]
         build_args = []
 
-        #if not cmake_generator:
+        # if not cmake_generator:
         #    cmake_args += ["-GNinja"]
 
-        self.parallel = 16
+        self.parallel = 4
         if "CMAKE_BUILD_PARALLEL_LEVEL" not in os.environ:
             if hasattr(self, "parallel") and self.parallel:
                 build_args += ["-j{}".format(self.parallel)]
@@ -49,6 +55,7 @@ class CMakeBuild(build_ext):
         subprocess.check_call(
             ["cmake", "--build", "."] + build_args, cwd=self.build_temp
         )
+
 
 setup(
     name="mplib",
@@ -68,15 +75,17 @@ setup(
         "Programming Language :: C++",
         "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
+        "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
         "Topic :: Education",
         "Topic :: Software Development :: Libraries :: Python Modules",
         "Topic :: Utilities",
     ],
-    python_requires='>=3.6',
-    install_requires=["numpy >= 1.20", "toppra >= 0.4.0", "transforms3d >= 0.3.1"],
+    python_requires=">=3.6",
+    install_requires=["toppra >= 0.4.0", "transforms3d >= 0.3.1"],
     ext_modules=[CMakeExtension("_mplib")],
     cmdclass={"build_ext": CMakeBuild},
     zip_safe=False,
     packages=["mplib"],
-    package_dir = {"mplib": "mplib/"}
+    package_dir={"mplib": "mplib"},
 )
